move translate under app
