+++ components/form-builder/app/translate/DownloadCSV.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 import { useTranslation } from "next-i18next";
-import { useTemplateStore } from "../store/useTemplateStore";
+import { useTemplateStore } from "../../store/useTemplateStore";
 import markdownToTxt from "markdown-to-txt";
-import { Button } from "../shared/Button";
+import { Button } from "../../shared/Button";
 
 const slugify = (str: string) =>
   str
