+++ components/form-builder/app/translate/RichText.tsx
@@ -1,9 +1,9 @@
 import { FormElement } from "@lib/types";
 import { useTranslation } from "next-i18next";
 import React from "react";
-import { RichTextEditor } from "../lexical-editor/RichTextEditor";
-import { useTemplateStore } from "../store/useTemplateStore";
-import { Language, LocalizedElementProperties } from "../types";
+import { RichTextEditor } from "../../lexical-editor/RichTextEditor";
+import { useTemplateStore } from "../../store/useTemplateStore";
+import { Language, LocalizedElementProperties } from "../../types";
 import { LanguageLabel } from "./LanguageLabel";
 
 export const RichText = ({
