+++ components/form-builder/app/translate/Translate.tsx
@@ -1,16 +1,16 @@
 import React from "react";
 import styled from "styled-components";
-import { useTemplateStore } from "../store/useTemplateStore";
+import { useTemplateStore } from "../../store/useTemplateStore";
 import { useTranslation } from "next-i18next";
 import { RichText } from "./RichText";
 import { SwapHoriz } from "@styled-icons/material/SwapHoriz";
 import { Title } from "./Title";
 import { Description } from "./Description";
 import { Options } from "./Options";
-import { LocalizedElementProperties } from "../types";
+import { LocalizedElementProperties } from "../../types";
 import { DownloadCSV } from "./DownloadCSV";
-import { RichTextEditor } from "../lexical-editor/RichTextEditor";
-import { Button } from "../shared/Button";
+import { RichTextEditor } from "../../lexical-editor/RichTextEditor";
+import { Button } from "../../shared/Button";
 import { LanguageLabel } from "./LanguageLabel";
 
 const LangSpan = styled.span`
